[refactor] extract run_cli() into src/runner.rs; pypi _main delegates to it because of PR 11
Problem: packages/pypi/src/lib.rs had a hand-rolled arg parser that only handled
'mille check'. Adding 'mille init' (or any future subcommand) required a manual
update to the pypi wrapper — a maintenance burden.

Fix:
- New src/runner.rs: contains run_cli() (pub) + all scan helpers (private)
  Any subcommand added to args.rs is automatically dispatched here.
- src/lib.rs: adds `pub mod runner`
- src/main.rs: reduced to `fn main() { mille::runner::run_cli(); }`
- packages/pypi/src/lib.rs: _main() now just calls mille_core::runner::run_cli()
  — zero lines to update when a new subcommand is added
