+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod domain;
 pub mod infrastructure;
 pub mod presentation;
+pub mod runner;
 pub mod usecase;
